Dash and rail
I tried a different way of creating the dash which was more successful.
It dashes for a limited timeframe  then has a cooldown before you can dash again.
Still buggy the dash is not a gradual speed boost more like teleportation.
I created a temporary rail model using blender but I could not get the collision or static body to work properly so that I can start on rail grinding.
created a single wall duplicating the ground and rotating it.