Very large change to the main scene, with buttons to change weapons and spawn enemies.
Added some buttons to main scene to change weapon to an assault rifle and shotgun
respectively. Added an additional button to spawn enemies at a designated SpawnPoint.
Also added a secondary light source for ease of testing.

Created a SpawnPoint scene which is an invisible Node3D that houses a script to
accept a signal from a button (or other things later on). Currently spawns 1
enemy.

Player gained an additional function to accept the button signal (and send to GunSlot).

Renamed interactRay to InteractRay for file naming conventions.